Add new qir_to_pytket method.
