+++ src/pytket_qirpass/__init__.py
@@ -1,5 +1,6 @@
 """Module for optimizing QIR using pytket."""
 
-__version__ = "0.8.0"
+__version__ = "0.9.0"
 
-from .apply_qirpass import apply_qirpass
+from .apply_qirpass import apply_qirpass as apply_qirpass
+from .qir_to_pytket import qir_to_pytket as qir_to_pytket
