+++ test/__init__.py
@@ -1,1 +1,1 @@
-import pytket_qirpass
+import pytket_qirpass as pytket_qirpass
